fix: ensure consistent hook usage and optimize state handling
