+++ .cspell.json
@@ -201,6 +201,8 @@
 		"longpress",
 		"Lorem",
 		"lucide",
+		"livekit",
+		"livekitroom",
 		"mappagination",
 		"mathieudutour",
 		"Mazen",
