+++ apps/web/.env
@@ -58,7 +58,7 @@ NEXT_PUBLIC_TWITTER_APP_NAME=ever-twitter
 TWITTER_CLIENT_ID=
 TWITTER_CLIENT_SECRET=
 
-#Livekit configiuration
+#Livekit configuration
 LIVEKIT_API_SECRET=
 LIVEKIT_API_KEY=
 NEXT_PUBLIC_LIVEKIT_URL=
